backend/requirements.txt
flask
flask-cors
Pillow
PyPDF2
img2pdf
werkzeug
